+++ CHANGELOG.md
@@ -4,8 +4,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+Please also update Silent Gear to 1.5.8+, contains armor color fixes
 ### Changed
 - (API) Replace Lombok getters with normal getters, removed some less useful getters
+### Fixed
+- Some armor models having incorrect textures (update Silent Gear as well for full fix) [#454]
 
 ## [3.5.1] - 2020-02-19
 ### Added