+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Chaos slime and wisp spawners checking the incorrect chunk, should fix [#445] and [#450]
+
+## [3.5.2] - 2020-04-15
 Please also update Silent Gear to 1.5.8+, contains armor color fixes
 ### Added
 - Configs to control chaos ore and ender ore spawns